Splitting api.c up. Future changes will also start to break out api.h.
